move allocated buffer class out of type.h
