+++ mooncake-store/src/storage_backend.cpp
@@ -1,6 +1,5 @@
-
-// #include "storage_backend.h"
 #include "storage_backend.h"
+
 #include <fcntl.h>
 #include <unistd.h>
 #include <sys/uio.h>
